+++ src/domains/edit/components/BaseEditor.tsx
